async-io: move FuseDevTask to session.rs
The FuseDevTask is similiar to FuseChannel, so move it into
session.rs.
